requirements.txt
apache-airflow==2.9.1
faker
minio
psycopg2-binary

